pyproject.toml
[project]
name = "oneyear"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.0",
    "pillow-heif>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
